DAM - COBOL - Final Project
In this project we were asked to create an application capable of creating destinations, making a reservation and deleting a reservation. All this I have done thanks to guide me through my teacher's code where he created a soccer team.

For 3 months I've been learning COBOl, it's not so bad, right?